chore: unify cigar stringification
